requirements/core.txt
pykube>=0.15.0
PyYAML>=3.12
